// src/data/types/home.types.ts
export enum CONTAINERS {
  ORDERS_THIS_YEAR = 'Orders This Year',
  TOTAL_REVENUE = 'Total Revenue',
  NEW_CUSTOMERS = 'New Customers',
  AVG_ORDERS_VALUE = 'Avg Order Value',
  CANCELED_ORDERS = 'Canceled Orders',
  CURRENT_MONTH_ORDERS = 'Current Month Orders',
  TOP_SOLD_PRODUCTS = 'Top Sold Products',
  CUSTOMER_GROWTH = 'Customer Growth',
  RECENT_ORDERS = 'Recent Orders',
  TOP_CUSTOMERS = 'Top Customers',
}